requirements.txt
streamlit==1.10.0
openai==0.28.0
python-dotenv==1.0.1
pydantic==2.7.4
fastapi==0.115.7
pandas==1.5.3